backend completed -v1.0

+++ backend/routes/users.routes.js
@@ -4,10 +4,14 @@ const { UserModel } = require("../models/users.models");
 const jwt = require("jsonwebtoken");
 const { blacklist } = require("../blacklist");
 const multer = require("multer");
+const { auth } = require("../middlewares/users.middleware");
 
 const userRouter = express.Router();
 
 //registration
+// Access: all
+// EndPoint: /user/register;
+// FRONTEND: when user/admin/teacher want to register in site;
 userRouter.post("/register", async (req, res) => {
   const { name, email, password, age, city, job, image } = req.body;
   const registeredUser = await UserModel.findOne({ email });
@@ -40,7 +44,11 @@ userRouter.post("/register", async (req, res) => {
   }
 });
 
-//get
+// login for users;
+// Access: All;
+// EndPoint: /users/login;
+// FRONTEND: when Admin/user/teacher want to login
+
 userRouter.post("/login", async (req, res) => {
   const { email, password } = req.body;
   try {
@@ -48,7 +56,7 @@ userRouter.post("/login", async (req, res) => {
     if (user) {
       bcrypt.compare(password, user.password, (err, result) => {
         // result == true
-        const token = jwt.sign({ userId: user._id, user: user.name, }, "arivu", {
+        const token = jwt.sign({ userId: user._id, user: user.name, role: user.role }, "arivu", {
           expiresIn: "7d",
         });
         const rToken = jwt.sign(
@@ -73,6 +81,9 @@ userRouter.post("/login", async (req, res) => {
 });
 
 //updation
+// Access: All
+// EndPoint: /users/update/:userId;
+// FRONTEND: when user want to update his information;
 userRouter.patch("/update/:userId", async (req, res) => {
   const { userId } = req.params;
   const payload = req.body;
@@ -88,15 +99,25 @@ userRouter.patch("/update/:userId", async (req, res) => {
 });
 
 
-//delete
-userRouter.delete("/delete/:userId", async (req, res) => {
-  const { userId } = req.params;
+//delete the user ;
+// Access: Admin 
+// EndPoint: /users/delete/:userId;
+// FRONTEND: when admin want to delete user/teacher
+userRouter.delete("/delete/:userId", auth, async (req, res) => {
+  
   try {
-    const deletedUser = await UserModel.find({ _id: userId });
-    await UserModel.findByIdAndDelete({ _id: userId });
-    res
-      .status(200)
-      .json({ msg: "user has been deleted", deletedUser: deletedUser });
+    if (req.body.role == "admin"){
+      const { userId } = req.params;
+      const deletedUser = await UserModel.find({ _id: userId });
+      await UserModel.findByIdAndDelete({ _id: userId });
+      res
+        .status(200)
+        .json({ msg: "user has been deleted", deletedUser: deletedUser });
+    }
+    else {
+      res.status(401).json({ error: "you don't have access to delete users" });
+    }
+    
   } catch (error) {
     res.status(400).json({ msg: error.message });
   }
@@ -104,6 +125,9 @@ userRouter.delete("/delete/:userId", async (req, res) => {
 
 
 //logout
+// Access: All 
+// EndPoint: /users/logout
+// FRONTEND: when users want to logout
 userRouter.get("/logout", (req, res) => {
   const token = req.headers.authorization?.split(" ")[1];
   try {
@@ -115,7 +139,11 @@ userRouter.get("/logout", (req, res) => {
 });
 
 
-// user course list 
+// list to courses user purchased
+// Access: All 
+// EndPoint: /users/userCourse/:userId
+// FRONTEND: When user want to see his purchased courses
+
 userRouter.get('/userCourse/:userId', async(req,res)=>{
   try{
       const userId = req.params.userId;
@@ -126,6 +154,38 @@ userRouter.get('/userCourse/:userId', async(req,res)=>{
       res.status(400).json({message:'Something Went Wrong',error:err.message})
   }
 })
+
+
+// add courseId to the user course array;
+// Access: All 
+// EndPoint: /users/addCourse/:courseId
+// FRONTEND: When user have purchased the couse and we have add it to the user course list;
+
+userRouter.post('/addCourse/:courseId',auth, async(req,res)=>{
+  try{
+    let id = req.body.userId;
+    const courseId = req.params.courseId
+// check is that course is already present or not;
+await UserModel.findOne({ _id: id, course: { $in: [courseId] } })
+  .then ( async (course) => {
+   //console.log(course);
+    if (course) {
+      res.status(400).json({error:'Course is already present for the user'});
+    } else {
+       let user =   await UserModel.findByIdAndUpdate(id,{$push:{course: courseId }})
+   res.status(201).json({message:'course added',user}) 
+    }
+  })
+  .catch(error => {
+    console.error('Error checking course:', error);
+  });
+  }catch(err){
+    res.status(400).json({message:'Something Went Wrong',error:err.message})
+  }
+})
+
+
+
 
 module.exports = {
   userRouter,
